fix(cli): extract converge report formatters under the 1200 LOC cap
converge.ts grew to 1213 lines on this branch (main is 1200). The
structural ratchet rejects that. Formatters move to converge-report.ts.
