+++ packages/remnic-cli/src/converge.ts
@@ -41,6 +41,7 @@ import { resolveAgentAccessAuthToken } from "@remnic/core/resolve-auth-token.js"
 import { parseFrontmatter } from "@remnic/core/storage.js";
 import type { ConvergeConflictPolicy } from "@remnic/core/types.js";
 import { loadConvergeIdentityCache, saveConvergeIdentityCache } from "./converge-identity-cache.js";
+import { formatConvergeApplyReport, formatConvergeReport } from "./converge-report.js";
 import {
   DEFAULT_PEER_REQUEST_TIMEOUT_MS,
   type PeerFileChunk,
@@ -975,42 +976,7 @@ async function updateCursorsForPlan(plan: ReconcilePlan, options: ConvergeApplyO
   }
 }
 
-export function formatConvergeReport(plan: ReconcilePlan): string {
-  const lines: string[] = [];
-  lines.push(`Convergence Status: ${plan.converged ? "CONVERGED" : "DIVERGED"}`);
-  lines.push("");
-  lines.push("Per-Namespace Summary:");
-  if (plan.byNamespace.length === 0) {
-    lines.push("  (no namespaces evaluated)");
-  } else {
-    for (const report of plan.byNamespace) {
-      lines.push(`  [${report.namespace}]`);
-      lines.push(`    identical:  ${report.identical}`);
-      lines.push(`    pull:       ${report.pull}`);
-      lines.push(`    push:       ${report.push}`);
-      lines.push(`    conflict:   ${report.conflict}`);
-      lines.push(`    suppress:   ${report.suppress}`);
-      lines.push(`    unresolved: ${report.unresolved}`);
-    }
-  }
-  return lines.join("\n");
-}
-
-export function formatConvergeApplyReport(result: ConvergeApplyResult): string {
-  const lines: string[] = [];
-  lines.push(`Convergence Execution Status: ${result.status.toUpperCase()}`);
-  lines.push(`Converged: ${result.converged ? "YES" : "NO"}`);
-  lines.push("");
-  lines.push("Transfers Executed:");
-  lines.push(`  pulled:             ${result.transfers.pulled}`);
-  lines.push(`  pushed:             ${result.transfers.pushed}`);
-  lines.push(`  conflictsResolved:  ${result.transfers.conflictsResolved}`);
-  lines.push(`  suppressed:         ${result.transfers.suppressed}`);
-  lines.push(`  failed:             ${result.transfers.failed}`);
-  lines.push("");
-  lines.push(formatConvergeReport(result.plan));
-  return lines.join("\n");
-}
+export { formatConvergeApplyReport, formatConvergeReport } from "./converge-report.js";
 
 /**
  * Convert the `--timeout <seconds>` flag to normalized milliseconds. The
